requirements.txt
openai-whisper>=1.0.0